Added server-client behaviour
Starting Timecode with --client argument from commandline will result
in starting as client.
A client is basically monitoring what's happening on the server, no remote
controlling possible.
The concept of a client is to give information to the staff, stagehands, stage
manager, FoH people or anyone, about what's happening.
This information consists of the current time, and custom information
provided from the server.